x86 Analysis [Capstone to Zydis]
